Refactor autolink subcommands into their own packages
