+++ pkg/cmd/repo/autolink/autolink.go
@@ -2,6 +2,7 @@ package autolink
 
 import (
 	"github.com/MakeNowJust/heredoc"
+	cmdList "github.com/cli/cli/v2/pkg/cmd/repo/autolink/list"
 	"github.com/cli/cli/v2/pkg/cmdutil"
 	"github.com/spf13/cobra"
 )
@@ -22,7 +23,7 @@ func NewCmdAutolink(f *cmdutil.Factory) *cobra.Command {
 	}
 	cmdutil.EnableRepoOverride(cmd, f)
 
-	cmd.AddCommand(newCmdList(f, nil))
+	cmd.AddCommand(cmdList.NewCmdList(f, nil))
 
 	return cmd
 }
