+++ pkg/cmd/repo/autolink/list/http_test.go
@@ -1,4 +1,4 @@
-package autolink
+package list
 
 import (
 	"fmt"
